+++ packages/nimble-components/src/banner/tests/banner.stories.ts
@@ -1,6 +1,6 @@
 import { html, when } from '@microsoft/fast-element';
 import { withActions } from '@storybook/addon-actions/decorator';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import { BannerSeverity } from '../types';
 import { bannerTag } from '..';
@@ -38,7 +38,7 @@ interface BannerArgs extends LabelUserArgs {
 
 const metadata: Meta<BannerArgs> = {
     title: 'Components/Banner',
-    decorators: [withActions],
+    decorators: [withActions<HtmlRenderer>],
     parameters: {
         actions: {
             handles: ['toggle']