+++ packages/nimble-components/src/select/tests/select.stories.ts
@@ -1,6 +1,6 @@
 import { html, repeat } from '@microsoft/fast-element';
 import { withActions } from '@storybook/addon-actions/decorator';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import {
     createUserSelectedThemeStory,
     disableStorybookZoomTransform
@@ -64,7 +64,7 @@ const optionSets = {
 
 const metadata: Meta<SelectArgs> = {
     title: 'Components/Select',
-    decorators: [withActions],
+    decorators: [withActions<HtmlRenderer>],
     parameters: {
         actions: {
             handles: ['change']