+++ packages/nimble-components/src/table-column/base/tests/table-column.stories.ts
@@ -1,5 +1,5 @@
 import { html, ref, when } from '@microsoft/fast-element';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { withActions } from '@storybook/addon-actions/decorator';
 import { createUserSelectedThemeStory } from '../../../utilities/tests/storybook';
 import {
@@ -79,7 +79,7 @@ information about specific types of column.`;
 
 const metadata: Meta<SharedTableArgs> = {
     title: 'Components/Table Column Configuration',
-    decorators: [withActions],
+    decorators: [withActions<HtmlRenderer>],
     parameters: {
         docs: {
             description: {