+++ packages/nimble-components/src/table-column/date-text/tests/table-column-date-text.stories.ts
@@ -1,5 +1,5 @@
 import { html, ref } from '@microsoft/fast-element';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { withActions } from '@storybook/addon-actions/decorator';
 import { createUserSelectedThemeStory } from '../../../utilities/tests/storybook';
 import { tableTag } from '../../../table';
@@ -55,7 +55,7 @@ const simpleData = [
 
 const metadata: Meta<SharedTableArgs> = {
     title: 'Components/Table Column: Date Text',
-    decorators: [withActions],
+    decorators: [withActions<HtmlRenderer>],
     parameters: {
         actions: {
             handles: sharedTableActions