+++ packages/nimble-components/src/table/tests/table-matrix.stories.ts
@@ -1,10 +1,8 @@
 import type { StoryFn, Meta } from '@storybook/html';
 import { html, ViewTemplate } from '@microsoft/fast-element';
+import { createStory } from '../../utilities/tests/storybook';
 import {
     createMatrixThemeStory,
-    createStory
-} from '../../utilities/tests/storybook';
-import {
     createMatrix,
     sharedMatrixParameters
 } from '../../utilities/tests/matrix';
@@ -17,6 +15,7 @@ import {
     TableRowSelectionMode
 } from '../types';
 import { tableColumnNumberTextTag } from '../../table-column/number-text';
+import { isChromatic } from '../../utilities/tests/isChromatic';
 
 const metadata: Meta = {
     title: 'Tests/Table',
@@ -104,7 +103,12 @@ const component = (
     [hierarchyStateName, hierarchyState]: HierarchyState
 ): ViewTemplate => html`
     <span>${() => `Selection mode: ${selectionMode ?? 'none'}, ${groupedStateName}, ${hierarchyStateName}`} </span>
-    <${tableTag} selection-mode="${() => selectionMode}"" id-field-name="id" parent-id-field-name="${() => (hierarchyState ? 'parentId' : undefined)}">
+    <${tableTag}
+        selection-mode="${() => selectionMode}"
+        id-field-name="id"
+        parent-id-field-name="${() => (hierarchyState ? 'parentId' : undefined)}"
+        style="${isChromatic() ? '--ni-private-spinner-animation-play-state:paused' : ''}"
+    >
         <${tableColumnTextTag} field-name="firstName" sort-direction="ascending" sort-index="0" group-index="${() => (groupedState ? '0' : undefined)}"><${iconUserTag}></${iconUserTag}></${tableColumnTextTag}>
         <${tableColumnTextTag} field-name="lastName">Last Name</${tableColumnTextTag}>
         <${tableColumnNumberTextTag} field-name="age" sort-direction="descending" sort-index="1" fractional-width=".5">Age</${tableColumnNumberTextTag}>
@@ -123,6 +127,13 @@ const playFunction = async (): Promise<void> => {
                         options: {
                             delayedHierarchyState:
                                 TableRecordDelayedHierarchyState.canLoadChildren
+                        }
+                    },
+                    {
+                        recordId: '1',
+                        options: {
+                            delayedHierarchyState:
+                                TableRecordDelayedHierarchyState.loadingChildren
                         }
                     }
                 ]);